+++ unify/src/components/Eventbutton/index.jsx
@@ -1,0 +1,13 @@
+import './style.css'
+
+function Eventbutton () {
+    return (
+        <div className="logo">
+              <button id="event-btn"><img src="./src/img/Addcircle.png" 
+              alt="EventBtn"/> Skapa event</button>
+            <img src="./src/img/ProfileIcon.png" alt="Profile"></img>
+            </div>
+    )
+}
+
+export default Eventbutton
