+++ unify/src/components/Eventbutton/style.css
@@ -1,0 +1,49 @@
+.logo {
+    display: flex;
+    gap: 1rem;
+    position: absolute;
+    top: 6.2rem;
+    right: .5rem;
+  }
+  
+  .logo img{
+    width: 26px;
+    height: 26px;
+    margin: 4px;
+  }
+  
+  #event-btn {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #157C79;
+    color: #fff;
+    font-family: HelveticaNeue;
+    font-size: 16px;
+    width: 125px;
+    height: 36px;
+    border-radius: 10px;
+    border: none;
+    cursor: pointer;
+    box-shadow: 1px 2px 2px #888888;
+  }
+  
+  #event-btn:hover {
+    background-color: #1b4848;
+  }
+  
+  #event-btn img {
+    margin: 0px 4px 0px 0px;
+  }
+
+  @media screen and (min-width: 960px) {
+    .event-btn {
+        display: flex;
+    }
+    .logo{
+        position: absolute;
+        top: 3rem;
+        right: 2rem;
+    }
+  }
+
