+++ unify/src/components/Togglebutton/style.css
@@ -1,0 +1,25 @@
+.toggle-button {
+    position: absolute;
+    top: 7rem;
+    left: 1rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    width: 30px;
+    height: 21px;
+  }
+  
+  .toggle-button .bar {
+    height: 3px;
+    width: 100%;
+    background-color: black;
+    border-radius: 10px;
+  }
+  @media screen and (min-width: 960px) {
+    .toggle-button {
+        display: none;
+        position: absolute;
+        top: 7.2rem;
+        left: 2rem;
+    }
+  }
