+++ unify/src/components/Unifybrand/index.jsx
@@ -1,0 +1,11 @@
+import './style.css'
+
+function Unifybrand () {
+    return (
+        <div className="brand-title">
+            <img src="./src/img/UnifyLogo.png" alt="UnifyIcon"/>
+        </div>
+    )
+}
+
+export default Unifybrand
