+++ unify/src/components/Unifybrand/style.css
@@ -1,0 +1,5 @@
+.brand-title {
+    font-size: 1.5rem;
+    margin: .5rem;
+    cursor: pointer;
+  }
