+++ unify/src/containers/Topnavbar/style.css
@@ -1,0 +1,17 @@
+.navbar {
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    align-items: center;
+    padding: 0rem  1rem;
+    background-color: #fff;
+    color: white;
+  }
+  
+  @media screen and (min-width: 960px) {
+    .navbar {
+        display: flex;
+        justify-content: flex-start;
+        flex-direction: row;
+    }
+  }
